requirements.txt
aniso8601==7.0.0
asgiref==3.7.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.6
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==4.3.0
django-filter==23.3
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
django-import-export==3.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
et-xmlfile==1.1.0
graphene==2.1.9
graphene-django==2.16.0
graphql-core==2.3.2
graphql-relay==2.0.1
MarkupPy==1.14
numpy==1.26.0
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.1.1
promise==2.3
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
Rx==1.6.3
singledispatch==4.1.0
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
text-unidecode==1.3
tzdata==2023.3
xlrd==2.0.1
xlwt==1.3.0
